requirements.txt
numpy==1.22.2
opencv_python==4.5.5.62
Pillow==9.1.1
python-dotenv==0.20.0
requests==2.27.1
torch==1.11.0
